Rename README to README.md.
In the next version a README file is introduced. Github will prefer to
show README.md if available and thus renaming the current README file to
README.md is a good idea.
This will preserve it and the purpose while still allowing the README
file to be present in the repo.